Get the URL authority when retrieving it from the HttpClient BaseAddress
When no explicit discovery address is specified, the authority is computed from the underlying HttpClient BaseAddress property.

Before this commit: the computed authority is the full URL.

After this commit: the computed authority is the authority part of the URL.

A HttpClient BaseAddress would typically include some base path to access an API. This base path is not part of the authority.